register

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 
@@ -14,12 +15,16 @@ class PostController extends Controller
      */
     public function index()
     {
+        if(!Auth::check()) {
+            return redirect('login');
+        }
+
         $posts = Post::active()->get();
         $view_data = [
             'posts' => $posts,            
         ];
 
-        return view('posts.index', $view_data);
+        return view('posts.index', $view_data);       
     }
 
     /**
@@ -27,6 +32,10 @@ class PostController extends Controller
      */
     public function create()
     {
+        if(!Auth::check()) {
+            return redirect('login');
+        }
+        
         return view('posts.create');
     }
 
@@ -35,6 +44,10 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        if(!Auth::check()) {
+            return redirect('login');
+        }
+        
         $title = $request->input('title');
         $content = $request->input('content');
 
@@ -56,6 +69,10 @@ class PostController extends Controller
      */
     public function show($id)
     {
+        if(!Auth::check()) {
+            return redirect('login');
+        }
+        
         $post = Post::where('id', $id)->first();
         $comments = $post->comments()->get();
         $total_comments = $post->total_comments();
@@ -73,6 +90,10 @@ class PostController extends Controller
      */
     public function edit($id)
     {
+        if(!Auth::check()) {
+            return redirect('login');
+        }
+        
         $post = Post::where('id', $id)->first();
         $view_data = [
             'post' => $post            
@@ -85,6 +106,10 @@ class PostController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!Auth::check()) {
+            return redirect('login');
+        }
+        
         $title = $request->input('title');
         $content = $request->input('content');
 
@@ -103,6 +128,10 @@ class PostController extends Controller
      */
     public function destroy($id)
     {
+        if(!Auth::check()) {
+            return redirect('login');
+        }
+        
         Post::where('id', $id)->delete();
                 
         return redirect("posts");
